requirements-dev.txt
pytest
pytest-playwright
pytest-mock
pytest-asyncio